requirements.txt
asgiref==3.4.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-import-export==2.7.0
et-xmlfile==1.1.0
MarkupPy==1.14
numpy==1.21.4
odfpy==1.4.1
openpyxl==3.0.9
pandas==1.3.4
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
xlrd==2.0.1
xlwt==1.3.0
